pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "kronos-hub"
version = "0.1.0"
description = "Integration-first umbrella project for ai-hedge-fund, TradingAgents, and Kronos."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "pydantic>=2.8.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.30.0",
]

[tool.setuptools.packages.find]
include = ["kronos_hub*", "apps*"]
